User story 4 User Roles completed

+++ app/controllers/wikis_controller.rb
@@ -1,63 +1,73 @@
-
 class WikisController < ApplicationController
+before_action :authenticate_user!
+  before_action :authenticate_user!, except: :index
+  after_action :verify_authorized, unless: :devise_controller?
+
 
   def index
     @wikis = Wiki.all
+    authorize @wikis
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
-    @wiki = Wiki.new
+    @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
-
-    @wiki.title = params[:wiki][:title]
-    @wiki.body = params[:wiki][:body]
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
        redirect_to @wiki
      else
-
        flash.now[:alert] = "There was an error saving the wiki. Please try again."
        render :new
      end
    end
 
+  def edit
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+  end
 
-     def edit
-       @wiki = Wiki.find(params[:id])
-     end
+  def update
+    @wiki = Wiki.find(params[:id])
+    @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
-     def update
-       @wiki = Wiki.find(params[:id])
-       @wiki.title = params[:wiki][:title]
-       @wiki.body = params[:wiki][:body]
+    if @wiki.save
+      flash[:notice] = "Wiki was updated."
+      redirect_to @wiki
+    else
+      flash.now[:alert] = "There was an error updating the wiki. Please try again."
+      render :edit
+    end
+  end
 
-       if @wiki.save
-         flash[:notice] = "Wiki was updated."
-         redirect_to @wiki
-       else
-         flash.now[:alert] = "There was an error updating the wiki. Please try again."
-         render :edit
-       end
-     end
+  def destroy
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
-     def destroy
-       @wiki = Wiki.find(params[:id])
+    if @wiki.destroy
+      flash[:notice] = "\"#{@wiki.title}\" was deleted."
+      redirect_to wikis_path
+    else
+      flash.now[:alert] = "There was an error deleting the post."
+      render :show
+    end
+  end
 
-       if @wiki.destroy
-         flash[:notice] = "\"#{@wiki.title}\" was deleted."
-         redirect_to wikis_path
-       else
-         flash.now[:alert] = "There was an error deleting the post."
-         render :show
-       end
-     end
-   end
+  private
+  
+  def wiki_params
+    params.require(:wiki).permit(:title, :body, :private)
+  end
+end
